requirements.txt
exceptiongroup>=1.1.3,<2.0.0
sniffio>=1.3.0,<2.0.0
packaging>=23.1,<28.0
certifi>=2023.7.22,<2028.0.0
Mako>=1.2.4,<2.0.0
aiofiles==23.2.1
aioshutil==1.3
gunicorn==21.2.0
SQLAlchemy[asyncio]==2.0.22
SQLAlchemy-Utils==0.41.1
psycopg[c]==3.1.12
# psycopg[binary]==3.1.12
alembic==1.12.0
# beans-logging[fastapi]==6.0.0
beans-logging-fastapi==1.0.2
onion-config[pydantic-v1]==5.1.0
aiohttp==3.8.6
fastapi[all]==0.99.1
